Ajout de la fonctionnalité de création de tickets avec intégration du formulaire dans le composant Test et gestion de l'affichage via un bouton.
